change default edge readInt to readByte

+++ computer-core/src/main/java/com/baidu/hugegraph/computer/core/graph/properties/DefaultProperties.java
@@ -79,7 +79,8 @@ public class DefaultProperties implements Properties {
     @Override
     public void read(RandomAccessInput in) throws IOException {
         this.keyValues.clear();
-        int size = in.readInt();
+        //int size = in.readInt();
+        int size = (int)in.readByte();
         for (int i = 0; i < size; i++) {
             String key = in.readUTF();
             ValueType valueType = SerialEnum.fromCode(ValueType.class,
